Removed URI scheme from jQuery link
When using this script on an https page, the script was blocked by default because it tried to load jQuery from an 'unsafe' locate (i.e. over http). By removing the URI scheme, the browser automatically applies the current page's scheme (i.e. http or https), and it works fine.